Fix byte-stream split decoding on big endian machines
